EY-2451 Flyttet sjekkliste til egen route
