+++ apps/etterlatte-saksbehandling-ui/client/src/shared/api/sjekkliste.ts
@@ -1,14 +1,13 @@
 import { apiClient } from '~shared/api/apiClient'
 import { ISjekkliste, ISjekklisteItem } from '~shared/types/Sjekkliste'
 
-export const hentSjekkliste = async (behandlingId: string) =>
-  apiClient.get<ISjekkliste>(`/behandling/${behandlingId}/sjekkliste`)
+export const hentSjekkliste = async (behandlingId: string) => apiClient.get<ISjekkliste>(`/sjekkliste/${behandlingId}`)
 
 export const opprettSjekkliste = async (behandlingId: string) =>
-  apiClient.post<ISjekkliste>(`/behandling/${behandlingId}/sjekkliste`, {})
+  apiClient.post<ISjekkliste>(`/sjekkliste/${behandlingId}`, {})
 
 export const oppdaterSjekkliste = async (sjekkliste: ISjekkliste) =>
-  apiClient.put<ISjekkliste>(`/behandling/${sjekkliste.id}/sjekkliste`, {
+  apiClient.put<ISjekkliste>(`/sjekkliste/${sjekkliste.id}`, {
     kommentar: sjekkliste.kommentar,
     adresseForBrev: sjekkliste.adresseForBrev,
     kontonrRegistrert: sjekkliste.kontonrRegistrert,
@@ -17,7 +16,7 @@ export const oppdaterSjekkliste = async (sjekkliste: ISjekkliste) =>
   })
 
 export const oppdaterSjekklisteItem = async (args: { behandlingId: string; item: ISjekklisteItem; checked: boolean }) =>
-  apiClient.post<ISjekklisteItem>(`/behandling/${args.behandlingId}/sjekkliste/${args.item.id}`, {
+  apiClient.post<ISjekklisteItem>(`/sjekkliste/${args.behandlingId}/item/${args.item.id}`, {
     avkrysset: args.checked,
     versjon: args.item.versjon,
   })
